fix: correct import path for DeleteCartEntity and refactor hotdeal logic in ShoppingCartService
